+++ rust/crates/astra-turn-core/Cargo.toml
@@ -17,12 +17,14 @@ chrono = { workspace = true }
 dashmap = "6.1.0"
 dirs = "6.0.0"
 fastrand = "2"
+futures-util = { workspace = true }
 regex = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 serde_yaml = "0.9.34"
 tokio = { workspace = true }
+tokio-util = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
 
 [dev-dependencies]
